Feature/create new security question and answer base component (#946)
* create security question base class, extend to reuse on registration flow

* fix hardcoded id's, loop and generate markup for form controls resulting in smaller markup

* extend setting back link, CTA label and save methods in child class

* uncomment route guard
