+++ src/app/features/create-account/create-account/create-account.component.html
@@ -22,7 +22,7 @@
             <label class="govuk-label" for="{{ item.name }}">
               {{ item.label }}
             </label>
-            <span *ngIf="submitted && form.get(item.name).errors" id="fullName-error" class="govuk-error-message">
+            <span *ngIf="submitted && form.get(item.name).errors" id="{{ item.name }}-error" class="govuk-error-message">
               <span class="govuk-visually-hidden">Error:</span> {{ getFirstErrorMessage(item.name) }}
             </span>
             <input
