+++ src/app/features/registration/security-question/security-question.component.html
@@ -13,51 +13,24 @@
 
         <p>You will need this if you forget your username or if you speak to Skills for Care about your account.</p>
 
-        <div class="govuk-form-group" [class.govuk-form-group--error]="submitted && getSecurityQuestion.errors">
-          <label class="govuk-label" for="securityQuestion">
-            Enter a security question
-          </label>
-          <span *ngIf="submitted && getSecurityQuestion.errors" id="securityQuestion-error" class="govuk-error-message">
-            <span class="govuk-visually-hidden">Error:</span>
-            <span *ngIf="getSecurityQuestion.hasError('required')">
-              {{ getFormErrorMessage('securityQuestion', 'required') }}
+        <ng-container *ngFor="let item of formControlsMap">
+          <div class="govuk-form-group" [class.govuk-form-group--error]="submitted && form.get(item.name).errors">
+            <label class="govuk-label" for="{{ item.name }}">
+              {{ item.label }}
+            </label>
+            <span *ngIf="submitted && form.get(item.name).errors" id="{{ item.name }}-error" class="govuk-error-message">
+              <span class="govuk-visually-hidden">Error:</span> {{ getFirstErrorMessage(item.name) }}
             </span>
-            <span *ngIf="getSecurityQuestion.hasError('maxlength')">
-              {{ getFormErrorMessage('securityQuestion', 'maxlength') }}
-            </span>
-          </span>
-          <input
-            class="govuk-input govuk-input--width-20"
-            formControlName="securityQuestion"
-            id="securityQuestion"
-            name="securityQuestion"
-            [class.govuk-input--error]="submitted && getSecurityQuestion.errors"
-            type="text"
-          />
-        </div>
-
-        <div class="govuk-form-group" [class.govuk-form-group--error]="submitted && getSecurityAnswer.errors">
-          <label class="govuk-label" for="securityAnswer">
-            Enter the answer to the security question
-          </label>
-          <span *ngIf="submitted && getSecurityAnswer.errors" id="securityAnswer-error" class="govuk-error-message">
-            <span class="govuk-visually-hidden">Error:</span>
-            <span *ngIf="getSecurityAnswer.hasError('required')">
-              {{ getFormErrorMessage('securityAnswer', 'required') }}
-            </span>
-            <span *ngIf="getSecurityAnswer.hasError('maxlength')">
-              {{ getFormErrorMessage('securityAnswer', 'maxlength') }}
-            </span>
-          </span>
-          <input
-            class="govuk-input govuk-input--width-20"
-            formControlName="securityAnswer"
-            id="securityAnswer"
-            name="securityAnswer"
-            [class.govuk-input--error]="submitted && getSecurityAnswer.errors"
-            type="text"
-          />
-        </div>
+            <input
+              class="govuk-input govuk-input--width-20"
+              [formControlName]="item.name"
+              id="{{ item.name }}"
+              name="{{ item.name }}"
+              [class.govuk-input--error]="submitted && form.get(item.name).errors"
+              type="text"
+            />
+          </div>
+        </ng-container>
       </fieldset>
 
       <button type="submit" class="govuk-button">{{ callToActionLabel }}</button>
